requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
python-multipart>=0.0.6
pydantic-settings>=2.0.0
supabase>=2.0.0
anthropic>=0.39.0
stripe>=10.0.0
mutagen>=1.47.0
firebase-admin>=6.4.0
apscheduler>=3.10.0
bcrypt>=4.0.0
pyjwt>=2.8.0
